recipe.apis.bookmark renamed as recipe.apis.bookmark_toggle : to make a clear distinction between Bookmark API & Bookmark Toggle API
