+++ app/recipes/apis/__init__.py
@@ -1,4 +1,4 @@
-from .bookmark import *
+from .bookmark_toggle import *
 from .like import *
 from .recipe import *
 from .recipe_name_choices import *
