+++ app/recipes/urls.py
@@ -2,7 +2,7 @@ from django.urls import path
 
 from recipes.apis.recipe_name_choices import RecipeNameChoicesListAPIView
 from .apis.validation import RecipeValidationAPIView
-from .apis.bookmark import BookmarkListCreateView
+from .apis.bookmark_toggle import BookmarkListCreateView
 from .apis.like import LikeListCreateView
 from .apis.recipe import RecipeListCreateView, RecipeRetrieveUpdateDestroyView
